try fix windows build warnings

+++ examples/ex009.cpp
@@ -11,7 +11,7 @@ exit ${RV}
 
 void ConvFixPoint(const std::string& pStr, int& pVal)
 {
-  pVal = roundf(100.0 * stof(pStr));
+  pVal = static_cast<int>(roundf(100.0f * std::stof(pStr)));
 }
 
 struct MyStruct
@@ -21,7 +21,7 @@ struct MyStruct
 
 void ConvMyStruct(const std::string& pStr, MyStruct& pVal)
 {
-  pVal.val = roundf(100.0 * stof(pStr));
+  pVal.val = static_cast<int>(roundf(100.0f * std::stof(pStr)));
 }
 
 int main()
@@ -31,7 +31,7 @@ int main()
   std::cout << "regular         = " << doc.GetCell<int>("Close", "2017-02-21") << "\n";
   std::cout << "fixpointfunc    = " << doc.GetCell<int>("Close", "2017-02-21", ConvFixPoint) << "\n";
 
-  auto convFixLambda = [](const std::string& pStr, int& pVal) { pVal = roundf(100.0 * stof(pStr)); };
+  auto convFixLambda = [](const std::string& pStr, int& pVal) { pVal = static_cast<int>(roundf(100.0f * stof(pStr))); };
   std::cout << "fixpointlambda  = " << doc.GetCell<int>("Close", "2017-02-21", convFixLambda) << "\n";
 
   std::cout << "mystruct        = " << doc.GetCell<MyStruct>("Close", "2017-02-21", ConvMyStruct).val << "\n";
